Revive Melon

+++ Source/Library/Common/types.h
@@ -12,7 +12,7 @@ typedef long long	s64int;
 typedef int 		s32int;
 typedef short		s16int;
 typedef char		s8int;
-typedef unsigned int size_t;
+typedef long unsigned int size_t;
 
 #define U64 unsigned long long
 #define S64 long long
